refactor(harness): extrair Hermes One model library (D-004) e revalidar cron
Move /api/model/library para hermes_one_model_library.py com mount fino
em web_server; atualiza patch-guard. Cron avaliacao-agente-dande voltou a ok.
